+++ .gitignore
@@ -1,3 +1,4 @@
 trade
 query
 logs/
+test/
