--- api/c++-1.7.0/config/apiConfig.ini
@@ -1,6 +1,0 @@
-[log]
-path=./apiLogs/
-level=ALL
-display=true
-monthdir=false
-limitsize=200
