+++ demo/linux/Makefile
@@ -1,14 +1,19 @@
 .PHONY: all
 
 all: hello-world \
+     trade-insert-order \
      query-account \
      query-account-asset \
      query-account-position \
      query-instrument \
-     query-combo
+     query-combo \
+     market-data
 
 hello-world: ./hello_world/trade.cc ./hello_world/trade.h
 	cd ./hello_world && make
+
+trade-insert-order: ./trade_insert_order/trade.cc ./trade_insert_order/trade.h
+	cd ./trade_insert_order && make
 
 query-account: ./query_account/query.cc ./query_account/query.h
 	cd ./query_account && make
@@ -24,3 +29,6 @@ query-instrument: ./query_instrument/query.cc ./query_instrument/query.h
 
 query-combo: ./query_combo/query.cc ./query_combo/query.h
 	cd ./query_combo && make
+
+market-data: ./market_data/market_data.cc ./market_data/market_data.h
+	cd ./market_data && make
