+++ demo/linux/market_data/MarketDataConfig.ini
@@ -1,0 +1,37 @@
+[Main]
+path=./tfApiLog/
+level = ALL
+display = false
+monthdir = true
+limitsize=10
+
+[RealTimeMarketData]
+RealTimeUnicastAddress=47.107.23.201
+RealTimeUnicastPort=13889
+
+
+StockAddress=230.10.10.1
+StockPort=23001
+
+IndexAddress=230.10.10.2
+IndexPort=23002
+
+OptionAddress=230.10.10.3
+OptionPort=23003
+
+FutureAddress=230.10.10.4
+FuturePort=23004
+
+TransactionAddress=230.10.10.5
+TransactionPort=23005
+
+OrderAddress=230.10.10.6
+OrderPort=23006
+
+[HistoryMarketData]
+HistoryUnicastAddress=47.106.45.9
+HistoryUnicastPort=2048
+
+[RealTimeIndicator]
+RealTimeIndicatorAddress=192.168.41.68
+RealTimeIndicatorPort=18239
