+++ demo/linux/query_account/Makefile
@@ -1,6 +1,5 @@
 CXX=g++
-CFLAGS=-std=c++11 -Wl,-rpath=../../../api/c++-1.7.0/lib/linux  -I. -I ../../../api/c++-1.7.0/include -L ../../../api/c++-1.7.0/lib/linux
-
+CFLAGS=-std=c++11 -Wl,-rpath=../../../api/TFTradeApi/uat/lib/linux  -I. -I ../../../api/TFTradeApi/uat/include -L ../../../api/TFTradeApi/uat/lib/linux
 
 query: query.cc query.h
 	$(CXX) $(CFLAGS) -o query query.cc -lTFTradeApi -lpthread
